Add TC3 automated test for new product creation and verification

+++ src/UI.Template/Components/Containers/EditProductContainer.cs
@@ -43,12 +43,12 @@ public class EditProductContainer(By locator) : BaseComponent(locator)
     /// </summary>
     /// <param name="value">The price to set.</param>
     /// <returns>True if the price was set correctly, false otherwise.</returns>
-    public bool SetPrice(int value)
-    {
-        Price.Clear();
-        Price.SendKeys(value.ToString(CultureInfo.InvariantCulture));
-        return int.Parse(Price.GetValue(), CultureInfo.InvariantCulture) == value;
-    }
+    public bool SetPrice(decimal value)
+{
+    Price.Clear();
+    Price.SendKeys(value.ToString(CultureInfo.InvariantCulture));
+    return decimal.Parse(Price.GetValue(), CultureInfo.InvariantCulture) == value;
+}
 
     /// <summary>
     /// Opens the basket container.
